init commit
adds -e and -d to encrypt and decrypt as well as an optional -o to name
the output file. for now, it's a one to one correspondence, so it only
encrypts or decrypts one file at a time.

we use xChaCha20Poly1305 to encrypt the file, then we use the same
algorithm with a unique nonce and a user-provided password which we use
to derive the second key to encrypt the key that was used to encrypt the
file. we store this encrypted copy of the key along with an argon2 hash
of the encrypted file's contents, the salt for the password, and the two
nonces, plus the original filename for the file we encrypted using this
key.

when the user wants to decrypt a file, we look at the metadata we saved
and try to find a stored hash that matches the new hash we generate of
the provided file's contents. when we find a match, we know which key we
need to use, and the user is prompted for their password.

still need to split the functionality up into more functions and do some
cleanup, but it's working currently.