add ls command and make some fixes
we need to use a different salt for the hash of the encrypted file
contents, and it's also helpful for the ls command to store the output
file name from when the file was encrypted, so the qlock file.

we also now check if the file already exists during encryption too

definitely some room to make things more DRY (Dont Repeat Yourself) but
it's okay for now.