add ability to pass multiple files, with numbered inputs for -o -n -p or comma separated lists
big overhaul to -p, -n, and -o during encryption/decryption so that we
can pass either an individual file, list of files, or a directory and
have each aforementioned flag provide either a comma separated list or
numbered inputs like -p1="" -p2="" -n1="" -n2="" etc and have them map
to each file provided in order, or alphabetically from root to child
directories in a deterministic way when the user provides a directory.

clean up some of the docs and the README to match the recent changes.

better password handling, use \, to include commas in passwords even
while using a comma separated list with -p